V1.0.3.2 Se cambian los botones y decoraciones a rounded-pill y se les añade shadow. Se rectifican pequeños bugs

+++ src/app/components/offerlist/offerlist.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { PRIMARY_OUTLET } from '@angular/router';
 import { AdminComponent } from 'src/app/pages/admin/admin.component';
 import { HomeComponent } from 'src/app/pages/home/home.component';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-offerslist',
@@ -31,7 +33,7 @@ export class OfferListComponent implements OnInit {
         this.arrayDataOffers = respose
       },
       error => {
-        alert(JSON.stringify(error))
+        alert(`Error ${error.status}: ${error.statusText}`)
       }
     )
   }
@@ -40,24 +42,46 @@ export class OfferListComponent implements OnInit {
     return this.arrayDataOffers
   }
 
-  fullOffer(id: string): void {
-    this._home.viewFullOffer(id)
-  }
+
 
   deleteOffer(id: string): void {
-    this._admin.deleteSelectedOffer(id).subscribe(
-      response => {
-        this._home.gotToHome()
-      },
-      error => {
-        alert(JSON.stringify(error))
 
+    Swal.fire({
+      title: '¿Estas seguro de eleminar la oferta?',
+      icon:'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      confirmButtonText: 'Si, borrar la oferta'
+    }).then(
+      response => {
+        this._admin.deleteSelectedOffer(id).subscribe(
+          response => {
+            Swal.fire('Oferta eliminada correctamente','success')
+            this.reeload()
+          },
+          error => {
+            alert(`Error ${error.status}: ${error.statusText}`)
+
+          }
+        )
       }
     )
   }
 
+  reeload(): void {
+    this._admin.roloadView()
+  }
+
   gotToNewOffer(): void {
     this._admin.gotToNewOffer()
+  }
+
+  fullOffer(id: string): void {
+    this._home.viewFullOffer(id)
+  }
+
+  fullOfferAdmin(id: string): void {
+    this._admin.viewFullOffer(id)
   }
 
   getModeAmin(): boolean {
